Finished

+++ deletePost.php
@@ -6,10 +6,11 @@ include 'dbconn.php';
 
 connect();
 $post_id = $_POST['p_id'];
+$user_id = $_POST['u_id'];
 echo $post_id.'<br>';
 
-$stmt = $conn->prepare("DELETE FROM POST WHERE POST_ID = ?");
-$stmt->bind_param("i",$post_id);
+$stmt = $conn->prepare("DELETE FROM POST WHERE POST_ID = ? AND USER_ID = ?");
+$stmt->bind_param("ii",$post_id,$user_id);
 
 if ($stmt->execute() === TRUE) {
 	// header('Location: '.$_SERVER['REQUEST_URI']);
